+++ elements/share/src/components/IntegrationButton.tsx
@@ -5,9 +5,10 @@ import styled from 'styled-components';
 import Button from '@atlaskit/button/custom-theme-button';
 import { CustomThemeButtonProps } from '@atlaskit/button/types';
 import { N500 } from '@atlaskit/theme/colors';
+import { token } from '@atlaskit/tokens';
 
 const IntegrationButtonCopyWrapper = styled.span`
-  color: ${(props) => props.theme.textColor || N500};
+  color: ${(props) => props.theme.textColor || token('color.text', N500)};
   display: flex;
   justify-content: left;
 `;