+++ elements/share/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @atlaskit/share
 
+## 3.1.0
+
+### Minor Changes
+
+- [`c3963026311`](https://bitbucket.org/atlassian/atlassian-frontend/commits/c3963026311) - [ux] Instrumented `@atlaskit/share` with the new theming package, `@atlaskit/tokens`.
+
+  New tokens will be visible only in applications configured to use the new Tokens API (currently in alpha).
+  These changes are intended to be interoperable with the legacy theme implementation. Legacy dark mode users should expect no visual or breaking changes.
+
 ## 3.0.13
 
 ### Patch Changes